Move tests to common file
